Validate user's 'name' before 'username'
The function username_from_ejson/1 is used to extract the username from
the user-provided JSON for a user.  That function checks for the 'name'
key followed by the 'username' key.

The validate_user_name/1 function was previously validating the
'username' key first, followed by the 'name' key if it didn't exist.
This mismatch meant that it was possible to upload a user item with an
invalid name that you would then be unable to delete.

This commit changes validate_user_name to validate the 'name' field
first.

Fixes #90

Signed-off-by: Steven Danna <steve@chef.io>